Move code for MATLAB generator from Ziggurat to example/ZigguratGenerator
